refactor to clean/clear
