+++ tramex-tools/src/interface/file_handler/mod.rs
@@ -1,0 +1,3 @@
+//! Websocket module
+
+pub mod file_handler;
