+++ tramex-tools/src/interface/websocket/mod.rs
@@ -1,9 +1,4 @@
 //! Websocket module
 
-pub mod layer;
-pub mod log_get;
-pub mod onelog;
-pub mod types;
-
 #[cfg(feature = "websocket")]
 pub mod ws_connection;
